[ADD]Panel admin y panel eventos

+++ TicketAPI/Dto/FeesResponseDto.cs
@@ -4,6 +4,6 @@
     {
         public Guid Id { get; set; }
         public string Type { get; set; }
-        public int Fee { get; set; }
+        public decimal Fee { get; set; }
     }
 }
